refactor(native): split chara_detail_recognizer into .h/.cpp
Second, larger step of the header/.cpp split, applied to the recognition logic
(previously 1101 lines, all implementation in the header, recompiled by every
consumer of native_api.cpp).

The header now keeps only what must stay there:
- The prediction structs (IndexPrediction, CharaPrediction, etc.) and the
  PredictionHistory declaration, since they are template arguments to
  recognizer::Model<> and appear as data members.
- The predict/predictWithConfidence function templates.
- Declarations of the nine recognizer classes and their members.

Everything else moves to the new .cpp verbatim: every recognizer method body, the
RaceBlockModelSet constructor, PredictionHistory::toJson, and the file-local
searchVertical helper (now in an anonymous namespace). This also fixes latent
leaky includes — the header previously used vlog_debug/log_warning without
including logger_util.h, relying on include order in native_api.cpp; the .cpp now
includes it explicitly.

Both consuming CMake source lists are updated in lockstep (native/CMakeLists.txt
and windows/runner/CMakeLists.txt).

Verified: umacapture_cli builds, umacapture_tests still passes (12/22),
`flutter build windows --debug` links the runner, and a stitch+recognize smoke
run on a copied fixture produced a valid record.json/prediction.json/trainee.jpg
(onCharaDetailFinished success:true, recognize path executing from the new .cpp).
